Custom tables with fpdf2
table_class.py and table_function.py perform create the same type of tables. The only difference is implementation. If you prefer using a function it is available, but I would recommend using the table_class.py to keep your code cleaner. 

This is an extrention of the fpdf2 library to provide more functionality when creating tables.